requirements.txt
Flask==3.1.0
Werkzeug==3.1.3
tensorflow==2.18.0
keras==3.6.0
numpy==2.0.2
Pillow==11.0.0
matplotlib==3.9.2
seaborn==0.13.2
scikit-learn==1.5.2